+++ src/mondayApi/queries/archiveGroup.ts
@@ -9,12 +9,12 @@ type ArchiveGroupType = {
 
 export const archiveGroup = async (
   client: MondayApi,
-  boardId: string,
+  boardId: number,
   groupId: string
 ): Promise<GroupType> =>
   (
     await client.api<ArchiveGroupType>(
-      `mutation archiveGroup($boardId: String!, $groupId: String!) {
+      `mutation archiveGroup($boardId: Int!, $groupId: String!) {
         archive_group (board_id: $boardId, group_id: $groupId) {
           ${GROUP_SUBQUERY}
         }