+++ src/mondayApi/queries/createGroup.ts
@@ -7,12 +7,12 @@ type CreateGroupType = {
 
 export const createGroup = async (
   client: MondayApi,
-  boardId: string,
+  boardId: number,
   groupName: string
 ): Promise<GroupType> =>
   (
     await client.api<CreateGroupType>(
-      `mutation createGroup($boardId: String!, $groupName: String!) {
+      `mutation createGroup($boardId: Int!, $groupName: String!) {
         create_group (board_id: $boardId, group_name: $groupName) {
           id
         }