+++ src/mondayApi/queries/updateColumn.ts
@@ -2,13 +2,13 @@ import MondayApi from "..";
 
 export const updateColumn = (
   client: MondayApi,
-  boardId: string,
+  boardId: number,
   itemId: number,
   columnId: string,
   value: string
 ): Promise<{ id: number }> =>
   client.api<{ id: number }>(
-    `mutation ($boardId: String!, $itemId: Int!, $columnId: String!, $value: String!) {
+    `mutation ($boardId: Int!, $itemId: Int!, $columnId: String!, $value: String!) {
       change_simple_column_value (board_id: $boardId, item_id: $itemId, column_id: $columnId, value: $value) {
         id
       }